Improve output of `install_node`
Now when `install_node` invokes `verify_install_node`, it
directs the stdout and stderr of the `verify_install_node`
subprocess to the stdio of the `install_node` process.

Previously, all of the output of `verify_install_node` could
only be viewd by the user when the `verify_install_node`
completed.