Delete `build.sh`
The `build.sh` file was to be used for building the base image
locally.  However, it is very similar to the `jenkins_build.sh`
file, and this similarity is causing some confusion on its use.

The `build.sh` file will first be deleted and future work can
look into local building that also incorporates future work that
is still work in progress.