Fix a failure to deploy if using yarn
This was caused by the fact that this repo's base image contains
a global install of `yarn`.

This installation of `yarn` is correct, and `yarn` can be used
in images that derive from the base image.

Further, if a global install of `yarn` is done during the building
of an image that derives from the base image in this repo, the
`yarn` installation will return an exit code indicating it did
succeed.  However, the `node_modules` dependencies for `yarn` will
not exist in the container.  As a result, `yarn` will not be able
to run.

This issue is actually not specific to `yarn`.  If, during the
build process of a Docker image, any package is installed globally
using `npm` and then installed globally again (in that image or a
derived image), the installation will not include the package's
dependencies.

The fix involves removing the global install of `yarn` in the
base image since it is nice to have, but not a strict requirement.