Remove manually running `node server.js`
Previously, if the `package.json` file did not contain a `start`
script, the application was started by calling `node server.js`.

Now, `npm start` or `yarn start` is always called to start the
application because those tools, themselves, already run
`node server.js` if a `start` script is not available in
`package.json`.