Test against installing npm@5.4.1
Previously, npm@2.7.1 was tested since when this version of npm was
installed, the following message would be displayed:

"fs: re-evaluating native module sources is not supported. If you
are using the graceful-fs module, please update it to a more recent
version."

However, this warning message is used to indicate that the module
being installed (npm) is using a version of graceful-fs that is
using a feature of Node that is no longer supported.

Since this is an issue with the npm, its dependencies, and the
features available in a specific version of Node, it is not specific
to the way npm is installed when building a Docker image in the
runtime builder.

Thus, a version of npm that is supported on Node 6 and 8 will be
tested since this PR only addresses installing versions of npm that
are supported on the installed version of Node.