Add support for custom workflows (#104)
Adds support for performing work that needs to be done to get the
application directory in a state ready for starting the
application.  Such work could include, for example, running gulp,
typescript, or webpack.

To use this functionality, supply a `gcp-build` script in your
`package.json` file.  If and only if this script is present, the runtime
builder will create a multi-stage Dockerfile to start your application.

In the first stage, your application contents are copied to the image,
dev dependencies are installed, and the `gcp-build` script is run.

The second stage uses the contents of the first stage except for the
`node_modules` directory.  Instead, non-dev dependencies are installed,
and the application is started.